requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.4.2
numpy==1.26.0
pytest==7.4.3
requests==2.31.0